requirements.txt
appdirs==1.4.4
attrs==25.3.0
Automat==25.4.16
beautifulsoup4==4.13.5
bs4==0.0.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
constantly==23.10.4
cryptography==45.0.7
cssselect==1.3.0
defusedxml==0.7.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fake-useragent==2.2.0
filelock==3.19.1
greenlet==3.2.4
h11==0.16.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
incremental==24.7.2
itemadapter==0.12.2
itemloaders==1.3.2
jmespath==1.0.1
lxml==6.0.1
numpy==2.2.6
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
parse==1.20.2
parsel==1.10.0
Protego==0.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyDispatcher==2.0.7
pyee==11.1.1
pyOpenSSL==25.1.0
pyquery==2.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
pywin32==311
queuelib==1.8.0
requests==2.32.5
requests-file==2.1.0
Scrapy==2.13.3
scrapyd==1.6.0
scrapyd-client==2.0.3
selenium==4.35.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
tldextract==5.3.0
tomli==2.2.1
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
Twisted==25.5.0
typing_extensions==4.14.1
tzdata==2025.2
uberegg==0.1.1
undetected-chromedriver==3.5.5
urllib3>=2.5.0,<3
w3lib==2.3.1
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.23.0
zope.interface==7.2
